+++ MAUICRUD/MVVM/ViewModels/ProdutoViewModel.cs
@@ -21,7 +21,7 @@ namespace MauiCrud.MVVM.ViewModels
         [ObservableProperty]
         private bool _isLoading;
         [ObservableProperty]
-        private int _code;
+        private long _code;
         [ObservableProperty] 
         private string _description = String.Empty;
         [ObservableProperty]
